Further development
git-svn-id: svn://localhost/EGE@383 1e37b749-e30f-da40-b4df-3e99ea2a9c57

+++ trunk/Sources/Airplay/TypesAirplay.h
@@ -3,14 +3,14 @@
 
 //--------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-typedef signed long         s32;            // i
-typedef unsigned long       u32;            // ui
+typedef signed int          s32;            // i
+typedef unsigned int        u32;            // ui
 typedef short               s16;            // si
 typedef unsigned short      u16;            // su
 typedef char                s8;             // c
 typedef unsigned char       u8;             // by
-typedef signed long         fp16;           // fp
-typedef signed long         fp8;            // fp
+typedef signed int          fp16;           // fp
+typedef signed int          fp8;            // fp
 typedef long long           s64;            // li
 typedef unsigned long long  u64;            // lu
 typedef float               float32;
